did stuff

+++ Project Andromeda/Assets/Scripts/Background.cs
@@ -5,7 +5,7 @@ using System;
 
 public class Background : MonoBehaviour {
 
-    public float speed;
+    public double speed;
 
     public Transform target;
     private SpawnObjects spawnObjects;
@@ -20,9 +20,8 @@ public class Background : MonoBehaviour {
     // Update is called once per frame
     void Update()
     {
-        float step = speed * Time.deltaTime;
+        float step = (float)speed * Time.deltaTime;
         transform.position = Vector2.MoveTowards(transform.position, target.position, step);
-
         if (Vector2.Distance(transform.position, target.position) < 0.1f)
         {
             transform.position = new Vector2(-26.6f, 0);
